refactor: isolate sse helpers
